[BUGFIX] Fix thread safety of write operations.
Was added two more interfaces in the store package.
  - KVReader is a read-only view of store;
  - KVWriter is a thread-safe API to write to store;

The KVStore interface now has methods to acquire readers
 and writers. The LevelDB implementation of store was modified
to the new API.

The leveldb Reader() method returns the default reader,
a singleton instance. The NewReader() returns a new fresh
instance of the reader.

The leveldb Writer() method returns a singleton writer. It's
not possible to create more than one writer at a time.